refactor(Validator): split function CarName
CarNameValidator to CarNameSplitValidator, CarNameValidator
